requirements.txt
aiohttp==3.7.3
APScheduler==3.7.0
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
idna==2.10
multidict==5.1.0
pytz==2020.5
requests==2.25.1
six==1.15.0
typing-extensions
tzlocal==2.1
urllib3==1.26.3
yarl==1.6.3
PyNaCl
hgtk
psycopg2
pillow
bs4
cloudscraper
py-cord